Reduce usage of the message 'artifact' dict
To help with https://pagure.io/fedora-ci/general/issue/436 ,
this reduces use of the 'artifact' dict from the Bodhi 'update
ready for testing' message to only use the 'task_id' and 'id'
(which is the build id) values from the 'build' dicts. This will
let us simplify the message format in Bodhi.

Signed-off-by: Adam Williamson <awilliam@redhat.com>